+++ packages/salesforce-adapter/src/constants.ts
@@ -372,6 +372,7 @@ export const CUSTOM_METADATA = 'CustomMetadata'
 export const FLOW_DEFINITION_METADATA_TYPE = 'FlowDefinition'
 export const INSTALLED_PACKAGE_METADATA = 'InstalledPackage'
 export const ACCOUNT_SETTINGS_METADATA_TYPE = 'AccountSettings'
+export const PERMISSION_SET_TYPE_ID_METADATA_TYPE = 'PermissionSet'
 
 // Artifitial Types
 export const CURRENCY_CODE_TYPE_NAME = 'CurrencyIsoCodes'
